mildoc_wxkf/requirements.txt
Flask==3.1.1
pycryptodome==3.23.0
requests==2.32.4
python-dotenv==1.1.1


# langchain相关依赖
langchain==0.3.26
langchain-openai==0.3.28
langchain-milvus==0.2.1
langchain-community==0.3.27

